+++ .vscode/settings.json
@@ -3,7 +3,7 @@
   "cSpell.words": [
     "Btns",
     "Faceoff",
-    "Gamefication",
+    "GamePlay",
     "Gamemode",
     "Leaderboard"
   ]
