+++ src/components/pages/GamePlay/Footer.js
