+++ src/components/pages/GamePlay/GameDrawStep.js
@@ -1,7 +1,7 @@
 //** Import Modules */
 import React, { useEffect, useState } from 'react';
-import { gsap } from 'gsap';
 import { useHistory } from 'react-router-dom';
+import { gsap } from 'gsap';
 import { Button } from 'antd';
 
 //** Import Components */
@@ -10,7 +10,7 @@ import UploadDocs from '../../common/UploadDocs';
 //** Import Assets */
 import boyImg from '../../../assets/images/gamemodeimg/LightingKid.png';
 
-export default function GameWriteStep(props) {
+export default function GameDrawStep(props) {
   // Get the history object
   const history = useHistory();
 
@@ -33,56 +33,57 @@ export default function GameWriteStep(props) {
     const triggerSubmitTimer = 1500;
 
     setTimeout(() => {
-      props.updateSubmissionData('HasWritten', true);
-      props.updateFileSubmissionData('writings', fileList);
-      props.updateBattleReady(true);
+      props.updateSubmissionData('HasDrawn', true);
+      props.updateFileSubmissionData('drawings', fileList);
+      props.updateCurStep('write');
 
       // Enable the modal window
-      const modalData = {
-        title: 'Get ready for the boss battle!',
-        description:
-          'To complete this mission, your work will be put up head-to-head against a boss. If you are ready to fight, submit your work!',
-        buttonTxt: "Let's Go!",
-      };
+      if (!props.modalClosed.draw) {
+        const modalData = {
+          title: 'Scribble a side quest!',
+          description:
+            'Grab your favorite pencil and some loose leaf sheets of paper. Based on the prompt at the end of the reading, scribble down a side quest by hand. When your story is complete, snap a photo of your pages and upload them.',
+          buttonTxt: "Let's Go!",
+        };
 
-      props.enableModalWindow(modalData);
+        props.enableModalWindow(modalData);
+        props.updateModalStatus('draw', true);
+      }
 
-      setIsUploading(false);
+      history.push(`${props.baseURL}/write`);
     }, triggerSubmitTimer);
   };
 
-  // This function handles when we make a full submission of the entire mission(after reading, drawing, and writing)
-  const handleFullSubmission = () => {};
-
   // This handles when we skip the drawing phase
   const handleSkip = () => {
-    props.updateSubmissionData('HasWritten', false);
-    props.updateFileSubmissionData('writings', []);
+    props.updateSubmissionData('HasDrawn', false);
+    props.updateFileSubmissionData('drawings', []);
 
-    if (!props.submissionData.HasDrawn) {
-      props.updateCurStep('draw');
-      props.updateBattleReady(false);
+    props.updateCurStep('write');
 
-      history.push(`${props.baseURL}/draw`);
-    } else {
-      props.updateCurStep('write');
-      props.updateBattleReady(true);
-
-      // Enable the modal window
+    // Enable the modal window
+    if (!props.modalClosed.draw) {
       const modalData = {
-        title: 'Get ready for the boss battle!',
+        title: 'Scribble a side quest!',
         description:
-          'To complete this mission, your work will be put up head-to-head against a boss. If you are ready to fight, submit your work!',
+          'Grab your favorite pencil and some loose leaf sheets of paper. Based on the prompt at the end of the reading, scribble down a side quest by hand. When your story is complete, snap a photo of your pages and upload them.',
         buttonTxt: "Let's Go!",
       };
 
       props.enableModalWindow(modalData);
+      props.updateModalStatus('draw', true);
     }
+
+    if (props.fileSubmissionData.writings.length === 0) {
+      props.updateBattleReady(false);
+    }
+
+    history.push(`${props.baseURL}/write`);
   };
 
   // Add some animation
   useEffect(() => {
-    gsap.from('#write-step', {
+    gsap.from('#draw-step', {
       opacity: 0,
       y: 200,
       duration: 1,
@@ -95,19 +96,12 @@ export default function GameWriteStep(props) {
       delay: 1,
     });
 
-    const battleBtnAnim = gsap.timeline({ repeat: -1, repeatDelay: 1 });
-    battleBtnAnim.to('.battle-btn button', { scale: 0.8, duration: 0.5 });
-    battleBtnAnim.to('.battle-btn button', { scale: 1.2, duration: 0.3 });
-    battleBtnAnim.to('.battle-btn button', { scale: 1.05, duration: 0.2 });
-    battleBtnAnim.to('.battle-btn button', { scale: 1.07, duration: 0.2 });
-    battleBtnAnim.to('.battle-btn button', { scale: 1, duration: 0.4 });
-
-    setFileList(props.fileSubmissionData.writings);
-  }, [props.fileSubmissionData.writings]);
+    setFileList(props.fileSubmissionData.drawings);
+  }, []);
 
   return (
-    <div id="write-step">
-      <div id="write" className="gamefication-content">
+    <div id="draw-step">
+      <div id="draw" className="gameplay-content">
         <div className="inner-container">
           <img src={boyImg} alt="Boy Image" className="boy-img" />
 
@@ -116,8 +110,8 @@ export default function GameWriteStep(props) {
 
             {fileList.length === 0 ? (
               <p>
-                When your story is complete, snap a photo of your pages and
-                upload them.
+                When you’re finished drawing, snap a photo and upload your
+                masterpiece.
               </p>
             ) : (
               <p>Double check that you’re uploading the right photo!</p>
@@ -126,7 +120,7 @@ export default function GameWriteStep(props) {
             <UploadDocs
               listType="picture"
               uploadButtonClassname="upload-picture"
-              uploadButtonText="Upload Your Writing"
+              uploadButtonText="Upload Your Drawing"
               handleChangeExtra={handleChange}
               handleSubmit={handleSubmit}
               alternateHandleSubmission={true}
@@ -145,16 +139,8 @@ export default function GameWriteStep(props) {
             )}
 
             <button className="skip-btn" onClick={handleSkip}>
-              {props.submissionData.HasDrawn
-                ? 'Just drawing, no story'
-                : "I'd rather draw"}
+              I’d rather write
             </button>
-
-            {props.battleReady && (
-              <div className="battle-btn">
-                <button>Begin Battle!</button>
-              </div>
-            )}
           </div>
         </div>
       </div>
