+++ src/components/pages/GamePlay/GamePlayMission.js
@@ -10,7 +10,7 @@ import GameReadStep from './GameReadStep';
 import GameDrawStep from './GameDrawStep';
 import GameWriteStep from './GameWriteStep';
 
-export default function GameficationMission(props) {
+export default function GamificationMission(props) {
   // GEt the history obj
   const history = useHistory();
