+++ src/components/pages/GamePlay/index.js
@@ -1,0 +1,1 @@
+export { default as GamePlayMain } from './GamePlayMain';
