--- src/components/pages/Gamefication/index.js
@@ -1,1 +1,0 @@
-export { default as GameficationMain } from './GameficationMain';
