+++ src/index.js
@@ -52,8 +52,8 @@ import { WritingSub } from './components/pages/WritingSub';
 import { Leaderboard } from './components/pages/Leaderboard';
 import FaceoffReveal from './components/pages/Animations/FaceoffReveal';
 
-// Gamefication Components
-import { GameficationMain } from './components/pages/Gamefication/index';
+// GamePlay Components
+import { GamePlayMain } from './components/pages/GamePlay/index';
 import { JoinTheSquad } from './components/pages/JoinTheSquad';
 import { PointShare } from './components/pages/PointShare';
 import { MatchUp } from './components/pages/MatchUp';
@@ -112,7 +112,7 @@ function App() {
       <DevModeHeader component={DevModeHeader} />
       <Switch>
         <Route exact path="/gamemode" component={Gamemode} />
-        <Route path="/gamefication" component={GameficationMain} />
+        <Route path="/gameplay" component={GamePlayMain} />
         <Route path="/gamemode/single" component={GamemodeButton} />
         <Route path="/login" component={LandingPage} />
         {/* any of the routes you need secured should be registered as ProtectedRoutes */}
