+++ src/styles/less/GamemodeStyles.less
@@ -1,5 +1,5 @@
-//** Gamefication Styles
-#gamefication {
+//** Gamification Styles
+#gameplay {
   /* Header Styles */
   .hero {
     margin-bottom: 50px;
@@ -12,7 +12,7 @@
     padding: 0 2%;
   }
 
-  /* Buttons on the main gamefication page */
+  /* Buttons on the main gamification page */
   .main-btns {
     button {
       width: 100%;
@@ -106,7 +106,7 @@
   }
 
   /* Inside step content where the story, draw, or write content is */
-  .gamefication-content {
+  .gameplay-content {
     padding: 200px 0 50px;
   }
 
@@ -177,7 +177,7 @@
       background: #ffd854;
     }
 
-    .gamefication-content {
+    .gamification-content {
       padding-bottom: 0 !important;
     }
